[!] error while compiling:
it compiles but doesn't show the images when i have done the logic part
i made a class called Calculator as it is an object we are trying to make
that performs addition, subtraction, division , and multiplication as
well as clearing and cancelling all operations . Dagi try to the rest
errors and modify what ever you like to make it work.

try using setters and getters i created within the Calculator class. i created them but i don't get it when i should use them.